Refactor: Extract method to dedicated files
