+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/util/Glob.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2020 New Vector Ltd
+ * Copyright 2020 The Matrix.org Foundation C.I.C.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.matrix.android.sdk.internal.util
+
+internal fun String.hasSpecialGlobChar(): Boolean {
+    return contains("*") || contains("?")
+}
+
+// Very simple glob to regexp converter
+internal fun String.simpleGlobToRegExp(): String {
+    var out = "" // "^"
+    for (element in this) {
+        when (element) {
+            '*'  -> out += ".*"
+            '?'  -> out += '.'.toString()
+            '.'  -> out += "\\."
+            '\\' -> out += "\\\\"
+            else -> out += element
+        }
+    }
+    out += "" // '$'.toString()
+    return out
+}
